UI: Reorganize and simplify configuration schema
