--- monkey/monkey_island/cc/ui/src/services/configuration/definitions/exploitation.js
@@ -1,26 +1,0 @@
-import {exploitationOptionsConfigurationSchema} from './exploitation_options.js';
-
-export const exploitationConfigurationSchema = {
-  'title': 'Exploiters',
-  'type': 'object',
-  'description': 'Choose which exploiters the Monkey will attempt.',
-  'properties': {
-    'brute_force': {
-      'title': 'Brute force exploiters',
-      'type': 'array',
-      'uniqueItems': true,
-      'items': {
-        '$ref': '#/definitions/brute_force_classes'
-      }
-    },
-    'vulnerability': {
-      'title': 'Vulnerability Exploiters',
-      'type': 'array',
-      'uniqueItems': true,
-      'items': {
-        '$ref': '#/definitions/vulnerability_classes'
-      }
-    },
-    'options': exploitationOptionsConfigurationSchema
-  }
-};
