+++ monkey/monkey_island/cc/ui/src/services/configuration/exploitationOptions.js
@@ -1,4 +1,4 @@
-export const exploitationOptionsConfigurationSchema = {
+const EXPLOITATION_OPTIONS_CONFIGURATION_SCHEMA = {
   'title': 'Exploiters Options',
   'type': 'object',
   'properties': {
@@ -13,3 +13,4 @@ export const exploitationOptionsConfigurationSchema = {
     }
   }
 }
+export default EXPLOITATION_OPTIONS_CONFIGURATION_SCHEMA;
