+++ monkey/monkey_island/cc/ui/src/services/configuration/exploiterClasses.js
@@ -1,4 +1,4 @@
-export const bruteForceExploiters = {
+export const BRUTE_FORCE_EXPLOITERS = {
     'title': 'Brute force exploiters',
     'description': 'Click on exploiter to get more information about it.'
     + '\u26A0'
@@ -68,7 +68,7 @@ export const bruteForceExploiters = {
     ]
 }
 
-export const vulnerabilityExploiters = {
+export const VULNERABILITY_EXPLOITERS = {
   'title': 'Vulnerability exploiters',
   'description': 'Click on exploiter to get more information about it.' +
   '\u26A0 Note that using unsafe exploits may cause craches of the exploited ' +
