+++ monkey/monkey_island/cc/ui/src/services/configuration/plugins.js
