+++ monkey/monkey_island/cc/ui/src/services/configuration/propagation/exploitation.js
@@ -1,0 +1,24 @@
+import EXPLOITATION_OPTIONS_CONFIGURATION_SCHEMA from '../exploitationOptions.js';
+import {BRUTE_FORCE_EXPLOITERS, VULNERABILITY_EXPLOITERS} from '../exploiterClasses';
+
+const EXPLOITATION_CONFIGURATION_SCHEMA = {
+  'title': 'Exploiters',
+  'properties': {
+    'brute_force': {
+      'items': BRUTE_FORCE_EXPLOITERS,
+      'title': 'Brute force exploiters',
+      'type': 'array',
+      'uniqueItems': true,
+    },
+    'vulnerability': {
+      'items': VULNERABILITY_EXPLOITERS,
+      'title': 'Vulnerability Exploiters',
+      'type': 'array',
+      'uniqueItems': true,
+    },
+  },
+  'options': EXPLOITATION_OPTIONS_CONFIGURATION_SCHEMA,
+  'type': 'object'
+};
+
+export default EXPLOITATION_CONFIGURATION_SCHEMA;
